Reimplement the support for MCRegisterInfo.
The new implementation more or less reflects everything interesting in this
class, and does so less on an as-needed, ad-hoc basis. The intent is to extend
this sort of interface to other implementations (principally MCInstrInfo), but
I haven't yet sat down to look at the full support needed for those yet.