Let the executor execute several testcases on random assembly code.
The inputs in question were relocatable .o files, although with no relocations.
Obviously, we're not going to support relocations in this code.

Another interesting facet is that the STOKE assembler seems to be playing
slightly fast and loose with symbol names--it's not distinguishing .name and
name, which causes the regular gas calls to fail.