Add support for reading in LLVM IR files.
This does illustrate an issue in comparison for the move pseudo-functions--
Strata/STOKE is doing sub-register compares, which means the trivial testcase
uses are failing (boo). This requires some smartness in doing the actual
comparison checks, which I really don't want to do.

Note that the in/out registers for  LLVM functions are read using LLVM
attributes. Too bad the LLVM rust bindings I'm using appear to be sitting on the
pre-3.3 gunk for attributes. One more reason to force the update?